rename wal_version to wal_edit
