+++ CMakeLists.txt
@@ -607,7 +607,7 @@ set(SOURCES
         db/version_edit_handler.cc
         db/version_set.cc
         db/wal_manager.cc
-        db/wal_version.cc
+        db/wal_edit.cc
         db/write_batch.cc
         db/write_batch_base.cc
         db/write_controller.cc
@@ -1076,7 +1076,7 @@ if(WITH_TESTS)
         db/version_edit_test.cc
         db/version_set_test.cc
         db/wal_manager_test.cc
-        db/wal_version_test.cc
+        db/wal_edit_test.cc
         db/write_batch_test.cc
         db/write_callback_test.cc
         db/write_controller_test.cc
