+++ Makefile
@@ -1491,7 +1491,7 @@ compact_on_deletion_collector_test: $(OBJ_DIR)/utilities/table_properties_collec
 wal_manager_test: $(OBJ_DIR)/db/wal_manager_test.o $(TEST_LIBRARY) $(LIBRARY)
 	$(AM_LINK)
 
-wal_version_test: $(OBJ_DIR)/db/wal_version_test.o $(TEST_LIBRARY) $(LIBRARY)
+wal_edit_test: $(OBJ_DIR)/db/wal_edit_test.o $(TEST_LIBRARY) $(LIBRARY)
 	$(AM_LINK)
 
 dbformat_test: $(OBJ_DIR)/db/dbformat_test.o $(TEST_LIBRARY) $(LIBRARY)
