+++ TARGETS
@@ -182,8 +182,8 @@ cpp_library(
         "db/version_edit.cc",
         "db/version_edit_handler.cc",
         "db/version_set.cc",
+        "db/wal_edit.cc",
         "db/wal_manager.cc",
-        "db/wal_version.cc",
         "db/write_batch.cc",
         "db/write_batch_base.cc",
         "db/write_controller.cc",
