+++ db/wal_edit.h
